+++ python2verilog/optimizer/__init__.py
@@ -2,4 +2,5 @@
 Optimizer Functions
 """
 
+from .helpers import backwards_replace
 from .increase_work import IncreaseWorkPerClockCycle
