+++ tests/optimizer/test_optimizer.py
@@ -7,10 +7,7 @@ from typing import Any
 from python2verilog import ir
 from python2verilog.backend.verilog import CodeGen
 from python2verilog.backend.verilog.codegen import CaseBuilder
-from python2verilog.optimizer.increase_work import (
-    IncreaseWorkPerClockCycle,
-    backwards_replace,
-)
+from python2verilog.optimizer import IncreaseWorkPerClockCycle, backwards_replace
 
 
 class TestGraphApplyMapping(unittest.TestCase):
